+++ app/reducers/sensorDetail.js
@@ -1,6 +1,7 @@
 import {
     SELECT_SENSOR, RECEIVE_SENSOR, UPDATE_DETAIL
 } from '../actions';
+import {getProcessedProperty} from '../utils/getConfig';
 
 const defaultState = {id: null, datapoints: []};
 
@@ -28,16 +29,18 @@ const sensorDetail = (state:backendsState = defaultState, action) => {
 };
 
 //TODO: sum to average
-function groupBy(array, col, value) {
+function groupBy(array, col, value, processed) {
     let r = [], o = {};
     array.forEach(function (a) {
         if (!o[a[col]]) {
             o[a[col]] = {};
             o[a[col]][col] = a[col];
             o[a[col]][value] = 0;
+            o[a[col]][processed] = a[processed];
             r.push(o[a[col]]);
         }
         o[a[col]][value] += +a[value];
+        o[a[col]][processed] += +a[processed];
     });
     return r;
 }
@@ -51,7 +54,7 @@ function collectdata(data){
 
     //console.log(data[0]);
     for(let key in data[0].properties){
-        data[0].properties[key] = groupBy(data[0].properties[key], 'label', 'average');
+        data[0].properties[key] = groupBy(data[0].properties[key], 'label', 'average', getProcessedProperty());
     }
     return data[0].properties;
 }